+++ backend/src/main/java/wooteco/retrospective/application/attendance/AttendanceService.java
@@ -1,15 +1,14 @@
-package wooteco.retrospective.application;
+package wooteco.retrospective.application.attendance;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import wooteco.retrospective.infrastructure.dao.attendance.AttendanceDao;
-import wooteco.retrospective.infrastructure.dao.attendance.TimeDao;
-import wooteco.retrospective.dao.member.MemberDao;
 import wooteco.retrospective.domain.attendance.Attendance;
 import wooteco.retrospective.domain.attendance.Time;
 import wooteco.retrospective.domain.member.Member;
 import wooteco.retrospective.dto.AttendanceRequest;
+import wooteco.retrospective.infrastructure.dao.attendance.AttendanceDao;
+import wooteco.retrospective.infrastructure.dao.attendance.TimeDao;
+import wooteco.retrospective.infrastructure.dao.member.MemberDao;
 
 @Service
 public class AttendanceService {
@@ -29,18 +28,18 @@ public class AttendanceService {
         validateTime(attendanceRequest);
 
         Member member = memberDao.findById(attendanceRequest.getMemberId())
-            .orElseThrow(RuntimeException::new);
+                .orElseThrow(RuntimeException::new);
         Time time = timeDao.findById(attendanceRequest.getTimeId())
-            .orElseThrow(RuntimeException::new);
+                .orElseThrow(RuntimeException::new);
         Attendance attendance = new Attendance(attendanceRequest.getDate(), member, time);
 
         return attendanceDao.insert(attendance);
     }
 
     private void validateTime(AttendanceRequest attendanceRequest) {
-        if(attendanceDao.isExistSameTime(
-            attendanceRequest.getMemberId(),
-            attendanceRequest.getTimeId()
+        if (attendanceDao.isExistSameTime(
+                attendanceRequest.getMemberId(),
+                attendanceRequest.getTimeId()
         )) {
             throw new IllegalArgumentException("이미 등록된 시간입니다.");
         }
